:recycle: Refactored command functions and message functions to own files. Also added public transport directions and options for getlectures command
